+++ src/PRo3D.SPICE/CooTransformation.FSharp.fs
@@ -60,7 +60,6 @@ module CooTransformation =
                 let forward = M44d(rot)
                 Trafo3d(forward, forward.Inverse) |> Some
             else
-                printfn "could not get rot trafo for frame: %s" fromFrame
-                Trafo3d.Identity |> Some
+                None
         )